Implement XPC connections on top of dispatch_mach
Also includes a serialization overhaul to be more object-oriented (each XPC object class worries about its own (de)serialization). The new serialization code also includes Mach port serialization (which we didn't have before).

The "test" service is not really a test; it's more of a sample. You have to test it out manually.

+++ internal-include/xpc/objects/array.h
@@ -8,7 +8,7 @@ XPC_CLASS_DECL(array);
 
 struct xpc_array_s {
 	struct xpc_object_s base;
-	NSUInteger size;
+	unsigned long size; // not NSUInteger or size_t because it needs to be `unsigned long` in 32-bit builds as well
 	XPC_CLASS(object)** array;
 };
